renamed PoweBIClient to PowerBIClient. fixed getDatasetAsync to return correct type directly
